requirements.txt
av==11.0.0
einops
flashy>=0.0.1
ffmpy
ftfy
gensim
h5py
httpx
huggingface-hub
hydra-core>=1.1
hydra_colorlog
julius
librosa==0.10.1
numpy<2.0.0
omegaconf==2.3.0
opencv-python==4.8.1.78
pydub
PyYAML==6.0.1
requests
scipy
soundfile
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
torch==2.2.0
torchaudio==2.2.0
torchmetrics
torchvision==0.17.0
tqdm
transformers==4.37.0
xformers==0.0.24
zipp